feat: add `Note` model and basic saving support
Add `NoteApplication` class to provide the application initialization code.
Move Joda-Time initialization code to application constructor, instead of main
activity constructor, to avoid duplicate initializations.

Enable code minification in release mode.

Add action bar support to the `EditNoteActivity`, currently with only a dummy
"Save Note" button available.
Add "save" icon for the "Save Note" button.

Enable Android Studio's built-in support for data binding.

Organize dependencies under their respective sections.
Add RxJava 2 and Requery dependencies for Requery support.

Add data binding metadata for `EditNoteActivity`, to bind it to the `Note`
model.

Add library providing the `RecyclerView` class.
Add RxAndroid as a dependency.

Add code to handle binding the `EditNoteActivity` note representations to the
view.
Add code to save/edit the note in the data store as necessary.

Add code to check for existence of notes and populate the list if it is empty.
Add `NoteAdapter` to adapt `Note` items from the `ReactiveEntityStore`,
extending from `QueryRecyclerAdapter`.

Initialize the `ReactiveEntityStore` with a set of empty tables if necessary in
a `getData()` method.

Modify `MainActivity` XML to utilize a `RecyclerView` instead of a
`RelativeLayout`.

Add a `BindingHolder` class extending the `RecyclerView.ViewHolder` class, to
hold bindings.

Add a `CreateNotes` class to generate a set of pre-defined notes, intended for
populating the list if it becomes empty.

Add a `note_item` XML layout definition to represent a note in the adapted
`RecyclerView`.