wip: add note tagging functionality
Add `Tag` class containing a tag name, and a M-to-M reference to the `Note`
class.
Add a "view tags" button to the menu for the main activity.
Increment database schema version to 2.

Add `TagActivity` to provide an activity for viewing tags.
Add `TagActivity` layout definition.
Add `TagActivity` to the activity list in `AndroidManifest.xml`.
Add a `viewTags` handler for when the button to view the list of tags is
pressed.

Add `CreateTags` class to create default example tags.

Add `tag_item.xml` file containing the XML representation of an individual
`Tag`, which will be data-bound to create the `TagItemBinding` class.